feat: Restructure project to support multiple files
The project structure has been refactored to accommodate multiple files within a project. This change introduces a `ProjectFile` interface and updates the `Project` interface to store an array of `ProjectFile` objects instead of a single `code` string.

This enables the application to manage and render multiple code files, paving the way for more complex project structures and improved organization. The preview functionality has also been updated to handle and transpile multiple files.